Make `docstrings` module private and move `exceptions` to `utils.exceptions`
