+++ rsciio/_docstrings.py
