+++ rsciio/digitalsurf/_api.py
@@ -50,8 +50,8 @@ import warnings
 # DictionaryTreeBrowser class handles the fancy metadata dictionnaries
 # from hyperspy.misc.utils import DictionaryTreeBrowser
 
-from rsciio.docstrings import FILENAME_DOC, LAZY_DOC, RETURNS_DOC
-from rsciio.exceptions import MountainsMapFileError
+from rsciio._docstrings import FILENAME_DOC, LAZY_DOC, RETURNS_DOC
+from rsciio.utils.exceptions import MountainsMapFileError
 
 _logger = logging.getLogger(__name__)
