+++ rsciio/edax/_api.py
@@ -24,7 +24,7 @@ import os
 import logging
 import numpy as np
 
-from rsciio.docstrings import (
+from rsciio._docstrings import (
     FILENAME_DOC,
     LAZY_DOC,
     ENDIANESS_DOC,
