+++ rsciio/empad/_api.py
@@ -22,7 +22,7 @@ import xml.etree.ElementTree as ET
 import numpy as np
 import logging
 
-from rsciio.docstrings import FILENAME_DOC, LAZY_DOC, RETURNS_DOC
+from rsciio._docstrings import FILENAME_DOC, LAZY_DOC, RETURNS_DOC
 from rsciio.utils.tools import _UREG
 from rsciio.utils.tools import convert_xml_to_dict
