+++ rsciio/hspy/_api.py
@@ -23,7 +23,7 @@ from pathlib import Path
 import dask.array as da
 import h5py
 
-from rsciio.docstrings import (
+from rsciio._docstrings import (
     FILENAME_DOC,
     LAZY_DOC,
     RETURNS_DOC,
