+++ rsciio/impulse/_api.py
@@ -4,7 +4,7 @@ from datetime import datetime as dt
 import csv
 import logging
 
-from rsciio.docstrings import FILENAME_DOC, RETURNS_DOC
+from rsciio._docstrings import FILENAME_DOC, RETURNS_DOC
 
 
 _logger = logging.getLogger(__name__)
