+++ rsciio/jobinyvon/_api.py
@@ -29,7 +29,7 @@ from copy import deepcopy
 import numpy as np
 from numpy.polynomial.polynomial import polyfit
 
-from rsciio.docstrings import FILENAME_DOC, RETURNS_DOC
+from rsciio._docstrings import FILENAME_DOC, RETURNS_DOC
 
 
 _logger = logging.getLogger(__name__)
