+++ rsciio/mrc/_api.py
@@ -25,7 +25,7 @@ import logging
 
 import numpy as np
 
-from rsciio.docstrings import (
+from rsciio._docstrings import (
     FILENAME_DOC,
     LAZY_DOC,
     ENDIANESS_DOC,
