+++ rsciio/msa/_api.py
@@ -24,7 +24,7 @@ import warnings
 
 import numpy as np
 
-from rsciio.docstrings import (
+from rsciio._docstrings import (
     FILENAME_DOC,
     ENCODING_DOC,
     RETURNS_DOC,
