+++ rsciio/protochips/_api.py
@@ -23,7 +23,7 @@ from datetime import datetime as dt
 import warnings
 import logging
 
-from rsciio.docstrings import FILENAME_DOC, RETURNS_DOC
+from rsciio._docstrings import FILENAME_DOC, RETURNS_DOC
 
 
 _logger = logging.getLogger(__name__)
