+++ rsciio/ripple/_api.py
@@ -28,7 +28,7 @@ import logging
 
 import numpy as np
 
-from rsciio.docstrings import (
+from rsciio._docstrings import (
     FILENAME_DOC,
     LAZY_DOC,
     ENCODING_DOC,
