+++ rsciio/tvips/_api.py
@@ -30,7 +30,7 @@ from dask.diagnostics import ProgressBar
 import pint
 from numba import njit
 
-from rsciio.docstrings import (
+from rsciio._docstrings import (
     FILENAME_DOC,
     LAZY_DOC,
     RETURNS_DOC,
