+++ rsciio/utils/exceptions.py
